pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
